+++ src/form.js
@@ -1,38 +1,38 @@
-import React from 'react';
-import { render } from 'react-dom';
-import data from './data/structure';
-import { UserForm } from './lib';
+import React from "react";
+import { render } from "react-dom";
+import data from "./data/structure";
+import { UserForm } from "./lib";
 
 class Form extends React.Component {
   formEdit = {
-    firstName: 'somename',
-    lastName: 'somelst',
-    username: 'user',
-    email: 'somemail@somedomain.com',
-    password1: 'random33',
-    password2: 'random33',
-    country: 'España',
-    addressPostalcode: '08004',
-    addressCity: '',
-    addressPhone: '111222333',
-    addressPostalcodeInvoice: '08004',
-    addressCityInvoice: '',
-    addressPhoneInvoice: '',
+    firstName: "somename",
+    lastName: "somelst",
+    username: "user",
+    email: "somemail@somedomain.com",
+    password1: "random33",
+    password2: "random33",
+    country: "España",
+    addressPostalcode: "08004",
+    addressCity: "",
+    addressPhone: "111222333",
+    addressPostalcodeInvoice: "08004",
+    addressCityInvoice: "",
+    addressPhoneInvoice: "",
     copyright: true,
-    optionHome: 'option_1'
+    optionHome: "option_1"
   };
   submitData = formSubmitData => {
-    console.log('formSubmitData ', formSubmitData);
+    console.log("formSubmitData ", formSubmitData);
   };
   render() {
     const Validation = params => {
       const { valid, message_type } = params;
       return !valid ? (
-        <div style={{ color: 'tomato' }}>
+        <div style={{ color: "tomato" }}>
           {message_type ? message_type : `$_customInvalidMessage`}
         </div>
       ) : (
-        ''
+        ""
       );
     };
     const SubmitButton = () => (
@@ -52,6 +52,7 @@ class Form extends React.Component {
           Validation={Validation}
           SubmitButton={SubmitButton}
           SubmitButtonText={`Submit`}
+          autoComplete="off"
         />
       </div>
     );